feat: adds api scope and base structure and frontend scope
